Fix: tipagem atributos de produto-service

+++ microservice/produto-service/src/main/java/com/ecommerce/produto_service/model/DTO/ProdutoCompraResponse.java
@@ -7,6 +7,6 @@ public record ProdutoCompraResponse(
         String nome,
         String descricao,
         BigDecimal preco,
-        double quantidade
+        Integer quantidade
 ) {
 }
